requirements.txt
jinja2==3.1.2
pytest==7.4.3
coverage==7.3.2
python-lsp-server
pygls==1.1.1
psycopg2-binary
python-dotenv
